chore: fix imports after package refactor
